// apps/auth/src/account/api/representations.ts
export interface CredentialMetadataRepresentationMessage {
  key: string;
  parameters?: string[];
}

export interface CredentialMetadataRepresentation {
  infoMessage: CredentialMetadataRepresentationMessage;
  infoProperties: CredentialMetadataRepresentationMessage[];
  warningMessageTitle: CredentialMetadataRepresentationMessage;
  warningMessageDescription: CredentialMetadataRepresentationMessage;
  credential: CredentialRepresentation;
}
// ...
export interface CredentialRepresentation {
  id: string;
  type: string;
  userLabel: string;
  createdDate: number;
  secretData: string;
  credentialData: string;
  priority: number;
  value: string;
  temporary: boolean;
  /**
   * @deprecated
   */
  device: string;
  /**
   * @deprecated
   */
  hashedSaltedValue: string;
  /**
   * @deprecated
   */
  salt: string;
  /**
   * @deprecated
   */
  hashIterations: number;
  /**
   * @deprecated
   */
  counter: number;
  /**
   * @deprecated
   */
  algorithm: string;
  /**
   * @deprecated
   */
  digits: number;
  /**
   * @deprecated
   */
  period: number;
  /**
   * @deprecated
   */
  config: { [index: string]: string[] };
}

export interface CredentialTypeMetadata {
  type: string;
  displayName: string;
  helpText: string;
  iconCssClass: string;
  createAction: string;
  updateAction: string;
  removeable: boolean;
  category: "basic-authentication" | "two-factor" | "passwordless";
}

export interface CredentialContainer {
  type: string;
  category: string;
  displayName: string;
  helptext: string;
  iconCssClass: string;
  createAction: string;
  updateAction: string;
  removeable: boolean;
  userCredentialMetadatas: CredentialMetadataRepresentation[];
  metadata: CredentialTypeMetadata;
}